src/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.5
command-not-found==0.3
contourpy==1.3.0
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
filelock==3.16.1
Flask==2.3.2
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
gast==0.4.0
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.66.1
gunicorn==20.1.0
h5py==3.11.0
httplib2==0.20.2
huggingface-hub==0.26.1
idna==3.8
imageio==2.36.0
importlib-metadata==4.6.4
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.4.2
kagglehub==0.3.3
keras==3.6.0
keyring==23.5.0
kiwisolver==1.4.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
libclang==18.1.1
lime==0.2.0.1
lit==18.1.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.1
mdurl==0.1.2
ml-dtypes==0.4.0
more-itertools==8.10.0
mpmath==1.3.0
namex==0.0.8
netifaces==0.11.0
networkx==3.4.2
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.5.3.2
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.5.82
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.5.82
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.5.82
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.3.0.75
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.3.61
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.6.82
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.6.3.83
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.5.1.3
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==1.5.3
pillow==11.0.0
protobuf==4.25.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==5.4.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.0
rsa==4.9
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.2.2
scipy==1.14.1
seaborn==0.12.2
SecretStorage==3.3.1
six==1.16.0
sympy==1.13.3
systemd-python==234
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tokenizers==0.20.1
torch==2.0.1
tqdm==4.66.5
transformers==4.46.0
triton==2.0.0
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.2
wadllib==1.3.6
Werkzeug==3.0.4
wrapt==1.14.1
zipp==1.0.0
